feat : update full code login & register
