+++ package.json
@@ -13,7 +13,6 @@
     "@ant-design/icons": "^5.4.0",
     "@reduxjs/toolkit": "^2.2.7",
     "@tanstack/react-query": "^5.51.24",
-    "animate.css": "^4.1.1",
     "antd": "^5.20.2",
     "axios": "^1.7.4",
     "concurrently": "^8.2.2",
