+++ src/interFaces/common.types.ts
@@ -1,11 +1,12 @@
 
 export interface IUser {
-    // id: number;
-    // point: number;
-    // phone: string;
-    // gender?: string;
-    // name?: string;
-    // email?: string;
-    // avatar?: string;
-    token: string;
+    id?: number | string;
+    point?: number;
+    phone?: string;
+    gender?: string;
+    name?: string;
+    email?: string;
+    avatar?: string;
+    token?: string;
+    password? : string;
   }
