+++ src/interFaces/user.ts
@@ -1,0 +1,10 @@
+export interface Iuser{
+    id : number | string
+}
+export interface IFormUser{
+    name? : string,
+    email : string,
+    password : string,
+    password_confirmation? : string;
+    remember?: boolean
+}
