--- src/layout/admin/Login.tsx
@@ -1,92 +1,0 @@
-
-import 'antd/dist/reset.css';
-import { Link } from 'react-router-dom';
-import LoginForm from '../../hooks/admin/LoginForm';
-import 'animate.css';
-
-const Login = () => {
-    return (
-        <section className="bg-white">
-            <div className="grid grid-cols-1 lg:grid-cols-2">
-                <div className="flex items-center justify-center px-4 py-10 bg-white sm:px-6 lg:px-8 sm:py-16 lg:py-24 ">
-                    <div className="animate__animated animate__bounceInUp xl:w-full xl:max-w-sm 2xl:max-w-md xl:mx-auto">
-                        <h2 className="text-3xl font-bold leading-tight text-black sm:text-4xl">
-                            Đăng nhập
-                        </h2>
-                        <p className="mt-2 text-base text-gray-600">
-                            Bạn chưa có tài khoản?{" "}
-                            <Link
-                                to={`/register`}
-                                title=""
-                                className="font-medium text-blue-600 transition-all duration-200 hover:text-blue-700 hover:underline focus:text-blue-700"
-                            >
-                                Tạo tài khoản ngay!
-                            </Link>
-                        </p>
-                        <LoginForm />
-                        <div className="mt-3 space-y-3">
-                            <button
-                                type="button"
-                                className="relative inline-flex items-center justify-center w-full px-4 py-4 text-base font-semibold text-gray-700 transition-all duration-200 bg-white border-2 border-gray-200 rounded-md hover:bg-gray-100 focus:bg-gray-100 hover:text-black focus:text-black focus:outline-none"
-                            >
-                                <div className="absolute inset-y-0 left-0 p-4">
-                                    <svg
-                                        className="w-6 h-6 text-rose-500"
-                                        xmlns="http://www.w3.org/2000/svg"
-                                        viewBox="0 0 24 24"
-                                        fill="currentColor"
-                                    >
-                                        <path d="M20.283 10.356h-8.327v3.451h4.792c-.446 2.193-2.313 3.453-4.792 3.453a5.27 5.27 0 0 1-5.279-5.28 5.27 5.27 0 0 1 5.279-5.279c1.259 0 2.397.447 3.29 1.178l2.6-2.599c-1.584-1.381-3.615-2.233-5.89-2.233a8.908 8.908 0 0 0-8.934 8.934 8.907 8.907 0 0 0 8.934 8.934c4.467 0 8.529-3.249 8.529-8.934 0-.528-.081-1.097-.202-1.625z" />
-                                    </svg>
-                                </div>
-                                Đăng nhập với Google
-                            </button>
-                            <button
-                                type="button"
-                                className="relative inline-flex items-center justify-center w-full px-4 py-4 text-base font-semibold text-gray-700 transition-all duration-200 bg-white border-2 border-gray-200 rounded-md hover:bg-gray-100 focus:bg-gray-100 hover:text-black focus:text-black focus:outline-none"
-                            >
-                                <div className="absolute inset-y-0 left-0 p-4">
-                                    <svg
-                                        className="w-6 h-6 text-[#2563EB]"
-                                        xmlns="http://www.w3.org/2000/svg"
-                                        viewBox="0 0 24 24"
-                                        fill="currentColor"
-                                    >
-                                        <path d="M13.397 20.997v-8.196h2.765l.411-3.209h-3.176V7.548c0-.926.258-1.56 1.587-1.56h1.684V3.127A22.336 22.336 0 0 0 14.201 3c-2.444 0-4.122 1.492-4.122 4.231v2.355H7.332v3.209h2.753v8.202h3.312z" />
-                                    </svg>
-                                </div>
-                               Đăng nhập với Facebook
-                            </button>
-                        </div>
-                    </div>
-                </div>
-                <div className="flex items-center justify-center px-4 py-10 sm:py-16 lg:py-24 bg-gray-50 sm:px-6 lg:px-8">
-                    <div>
-                        <img
-                            className="w-full mx-auto"
-                            src="https://cdn.rareblocks.xyz/collection/celebration/images/signup/1/cards.png"
-                            alt=""
-                        />
-                        <div className="w-full max-w-md mx-auto xl:max-w-xl">
-                            <h3 className="text-2xl font-bold text-center text-black">
-                                Design your own card
-                            </h3>
-                            <p className="leading-relaxed text-center text-gray-500 mt-2.5">
-                                Amet minim mollit non deserunt ullamco est sit aliqua dolor do amet
-                                sint. Velit officia consequat duis.
-                            </p>
-                            <div className="flex items-center justify-center mt-10 space-x-3">
-                                <div className="bg-orange-500 rounded-full w-20 h-1.5" />
-                                <div className="bg-gray-200 rounded-full w-12 h-1.5" />
-                                <div className="bg-gray-200 rounded-full w-12 h-1.5" />
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </section>
-    );
-};
-
-export default Login;
-
