+++ src/pages/admin/Register/utils/RegisterForm.services.ts
@@ -1,0 +1,19 @@
+import { notification } from "antd";
+import { IFormUser } from "../../../../interFaces/user";
+import ApiUtils from "../../../../utils/api/api.utils";
+
+export const registerService = async (body: IFormUser) => {
+  try {
+    const res = await ApiUtils.post(
+      "/api/register",
+      body
+    );
+    return res;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  } catch (error : any) {
+    notification.error({
+      message: error.message,
+      description:error.errors[0],
+    });
+  }
+};
